avoid cyclic dep
