+++ src/main/scala-sbt-0.13/CompatExtras.scala
@@ -1,6 +1,5 @@
 package org.ensime
 
-import org.ensime.EnsimeExtrasPlugin.{fileInProject, noChanges}
 import sbt.Defaults.createTestRunners
 import sbt.Keys._
 import sbt.Tests.Execution
@@ -8,7 +7,7 @@ import sbt._
 import sbt.complete.{DefaultParsers, Parser}
 import sbt.testing.Framework
 
-object CompatExtrasKeys {
+trait CompatExtrasKeys {
   val ensimeTestOnlyDebug = inputKey[Unit](
     "The equivalent of ensimeRunDebug for testOnly command"
   )
@@ -16,13 +15,19 @@ object CompatExtrasKeys {
   val ensimeCompileOnly = inputKey[Unit](
     "Compiles a single scala file"
   )
+
+  val ensimeDebuggingArgs = settingKey[Seq[String]](
+    "Java args for for debugging"
+  )
+
 }
 
-trait CompatExtras {
-  import CompatExtrasKeys._
-  import EnsimeExtrasKeys._
+object CompatExtrasKeysHack extends CompatExtrasKeys
 
-  val compatSettings: Seq[Setting[_]] = Seq(
+trait CompatExtras {
+  import CompatExtrasKeysHack._
+
+  def compatSettings: Seq[Setting[_]] = Seq(
     ensimeTestOnlyDebug in Test := testOnlyWithSettings(
       Test,
       extraArgs = ensimeDebuggingArgs
@@ -160,5 +165,23 @@ trait CompatExtras {
       input, noChanges, cp.map(_.data) :+ out, out, opts,
       noopCallback, merrs, in.incSetup.cache, s.log
     )
+  }
+
+  private[ensime] val noChanges = new xsbti.compile.DependencyChanges {
+    def isEmpty = true
+    def modifiedBinaries = Array()
+    def modifiedClasses = Array()
+  }
+
+  private[ensime] def fileInProject(arg: String, sourceDirs: Seq[File]): File = {
+    val input = file(arg).getCanonicalFile
+    val here = sourceDirs.exists { dir => input.getPath.startsWith(dir.getPath) }
+    if (!here || !input.exists())
+      throw new IllegalArgumentException(s"$arg not associated to $sourceDirs")
+
+    if (!input.getName.endsWith(".scala"))
+      throw new IllegalArgumentException(s"only .scala files are supported: $arg")
+
+    input
   }
 }
