+++ src/main/scala-sbt-1.0/CompatExtras.scala
@@ -1,6 +1,12 @@
 package org.ensime
 
-import sbt.Setting
+import sbt._
+
+trait CompatExtrasKeys {
+  val ensimeDebuggingArgs = settingKey[Seq[String]](
+    "Java args for for debugging"
+  )
+}
 
 trait CompatExtras {
   val compatSettings: Seq[Setting[_]] = Nil
